# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Cogitorium
  version: "1"
  description: |-
    The HTTP surface of a Cogitorium install, generated from the server's own
    route table by a test that fails when the two disagree — so this lists every
    endpoint that exists, and only those.

    Request bodies are described where the handler decodes into a named type,
    and those schemas are the parser's own definition rather than a second
    account of it. The rest are being named route by route; until then their
    shape is in the reference at https://orkcom-tech.github.io/cogitorium/.
    A test holds the count of undescribed ones so it can only fall.

    Response bodies are not described yet.

    The version above is the API surface — everything here lives under /api/v1 —
    and deliberately not the build's version: a document that changed with every
    release would report a new API on a day nothing about it moved, and would
    fail its own generation test on any build that stamps a version in.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: http://127.0.0.1:8688
    description: the default listen address
components:
  securitySchemes:
    token:
      type: http
      scheme: bearer
      description: |-
        A user's token. On a loopback listen address a local request is treated
        as the admin without one; change the address and it becomes required.
    inletKey:
      type: http
      scheme: bearer
      description: |-
        A receiver's own key, which opens that receiver and nothing else. These
        are the only paths exempt from normal authentication.
security:
  - token: []
paths:
...
  "/api/v1/env/{name}":
    parameters:
      - name: "name"
        in: path
        required: true
        schema:
          type: string
    delete:
      summary: "Delete env/{name}"
      responses:
        default:
          description: see the reference
    put:
      summary: "Replace env/{name}"
      responses:
        default:
          description: see the reference
...
  "/api/v1/queue/{id}":
    parameters:
      - name: "id"
        in: path
        required: true
        schema:
          type: integer
          format: int64
    delete:
      summary: "Delete queue/{id}"
      responses:
        default:
          description: see the reference
...
  "/api/v1/workspaces":
    get:
      summary: "Read workspaces"
      responses:
        default:
          description: see the reference
    post:
      summary: "Create workspaces"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                "description":
                  type: string
                "name":
                  type: string
                "orchestrator_model_id":
                  type: integer
      responses:
        default:
          description: see the reference
...
  "/api/v1/workspaces/{id}/env":
    parameters:
      - name: "id"
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      summary: "Read workspaces/{id}/env"
      responses:
        default:
          description: see the reference
  "/api/v1/workspaces/{id}/env/{name}":
    parameters:
      - name: "id"
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: "name"
        in: path
        required: true
        schema:
          type: string
    delete:
      summary: "Delete workspaces/{id}/env/{name}"
      responses:
        default:
          description: see the reference
    put:
      summary: "Replace workspaces/{id}/env/{name}"
      responses:
        default:
          description: see the reference
...
  "/api/v1/workspaces/{id}/queue":
    parameters:
      - name: "id"
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      summary: "Read workspaces/{id}/queue"
      responses:
        default:
          description: see the reference
...